Moved db code to db.cpp and db.h
